backend/api/requirements.txt
# FastAPI REST API Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=0.19.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Data validation
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Security
bcrypt>=4.0.0

# Additional utilities
python-multipart>=0.0.6
